- Converted all in-house Linear Algebra operations (Vector, Matrix, Quaternion) into jME equivalents and changed all 'double' type position oeprations to 'float' types.
- Updated rendering procedure to separately light reverse mesh backfaces (or ignore if the backface culling render material configuration is enabled)

- Simplified exception logging format by removing timestamp logs from each exception entry; now, the runtime entry header contains the timestamp while the file header indicates the date

+++ src/main/java/com/transcendruins/assets/modelassets/attack/AttackInstance.java
@@ -12,9 +12,9 @@ public final class AttackInstance extends Instance {
         return damage;
     }
 
-    private double range;
+    private float range;
 
-    public final double getRange() {
+    public final float getRange() {
 
         return range;
     }
@@ -39,7 +39,7 @@ public final class AttackInstance extends Instance {
         AttackSchema attributes = (AttackSchema) attributeSet;
 
         damage = calculateAttribute(attributes.getDamage(), damage, attributes, 0);
-        range = calculateAttribute(attributes.getRange(), range, attributes, 0.0);
+        range = calculateAttribute(attributes.getRange(), range, attributes, 0.0f);
         speed = calculateAttribute(attributes.getSpeed(), speed, attributes, -1.0);
         cooldown = calculateAttribute(attributes.getCooldown(), cooldown, attributes, 0.0);
     }
